Repurpose DataPage as MarketPage
