+++ ViewModels/MainWindowViewModel.cs
@@ -48,10 +48,10 @@ namespace TractorMarket.ViewModels
                 },
                 new NavigationItem()
                 {
-                    Content = "Data",
-                    PageTag = "data",
-                    Icon = SymbolRegular.DataHistogram24,
-                    PageType = typeof(Views.Pages.DataPage)
+                    Content = "Markt",
+                    PageTag = "market",
+                    Icon = SymbolRegular.VehicleTruckBag24,
+                    PageType = typeof(Views.Pages.MarketPage)
                 }
             };
