+++ Views/Pages/MarketPage.xaml
@@ -1,0 +1,32 @@
+﻿<ui:UiPage
+    x:Class="TractorMarket.Views.Pages.MarketPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:local="clr-namespace:TractorMarket.Views.Pages"
+    xmlns:entities="clr-namespace:TractorMarket.Entities"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:ui="http://schemas.lepo.co/wpfui/2022/xaml"
+    Title="MarketPage"
+    d:DataContext="{d:DesignInstance local:MarketPage, IsDesignTimeCreatable=False}"
+    d:DesignHeight="450"
+    d:DesignWidth="800"
+    d:Background="{DynamicResource ApplicationBackgroundBrush}"
+    Foreground="{DynamicResource TextFillColorPrimaryBrush}"
+    mc:Ignorable="d">
+
+    <Grid Margin="56,0">
+      <ui:VirtualizingItemsControl
+        Foreground="{DynamicResource TextFillColorSecondaryBrush}"
+        ItemsSource="{Binding ViewModel.TractorsForCustomers, Mode=OneWay}"
+        VirtualizingPanel.CacheLengthUnit="Item"
+      >
+        <ItemsControl.ItemTemplate>
+          <DataTemplate DataType="{x:Type entities:Tractor}">
+            <ui:Card>
+            </ui:Card>
+          </DataTemplate>
+        </ItemsControl.ItemTemplate>
+      </ui:VirtualizingItemsControl>
+    </Grid>
+</ui:UiPage>
